api/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2
pydantic-settings>=2
orjson
SQLAlchemy>=2.0
psycopg2-binary>=2.9
alembic>=1.13
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
PyJWT>=2.9
python-multipart>=0.0.9
pydantic[email]>=2
